+++ java/src/org/openqa/selenium/chrome/ChromeDriverInfo.java
@@ -18,6 +18,8 @@
 package org.openqa.selenium.chrome;
 
 import com.google.auto.service.AutoService;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.ImmutableCapabilities;
@@ -43,6 +45,16 @@ public class ChromeDriverInfo extends ChromiumDriverInfo {
 
   @Override
   public Capabilities getCanonicalCapabilities() {
+    if (!"jdk-http-client".equalsIgnoreCase(System.getProperty("webdriver.http.factory", ""))) {
+      // Allowing any origin "*" through remote-allow-origins might sound risky but an attacker
+      // would need to know the port used to start DevTools to establish a connection. Given
+      // these sessions are relatively short-lived, the risk is reduced. Only set when the Java
+      // 11 client is not used.
+      return new ImmutableCapabilities(
+        CapabilityType.BROWSER_NAME, CHROME.browserName(),
+        ChromeOptions.CAPABILITY,
+        ImmutableMap.of("args", ImmutableList.of("--remote-allow-origins=*")));
+    }
     return new ImmutableCapabilities(CapabilityType.BROWSER_NAME, CHROME.browserName());
   }
 
@@ -65,7 +77,7 @@ public class ChromeDriverInfo extends ChromiumDriverInfo {
   @Override
   public boolean isAvailable() {
     try {
-      DriverFinder.getPath(ChromeDriverService.createDefaultService());
+      DriverFinder.getPath(ChromeDriverService.createDefaultService(), getCanonicalCapabilities());
       return true;
     } catch (IllegalStateException | WebDriverException e) {
       return false;