+++ java/src/org/openqa/selenium/chromium/ChromiumOptions.java
@@ -32,6 +32,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.stream.Stream;
@@ -74,6 +75,13 @@ public class ChromiumOptions<T extends ChromiumOptions<?>> extends AbstractDrive
   public ChromiumOptions(String capabilityType, String browserType, String capability) {
     this.capabilityName = capability;
     setCapability(capabilityType, browserType);
+    if (!"jdk-http-client".equalsIgnoreCase(System.getProperty("webdriver.http.factory", ""))) {
+      // Allowing any origin "*" might sound risky but an attacker would need to know
+      // the port used to start DevTools to establish a connection. Given these sessions
+      // are relatively short-lived, the risk is reduced. Only set when the Java 11 client
+      // is not used.
+      addArguments("--remote-allow-origins=*");
+    }
   }
 
   /**
@@ -113,7 +121,7 @@ public class ChromiumOptions<T extends ChromiumOptions<?>> extends AbstractDrive
    * Adds additional command line arguments to be used when starting Chrome.
    * For example:
    * <pre><code>
-   *   options.setArguments(
+   *   options.addArguments(
    *       "load-extension=/path/to/unpacked_extension",
    *       "allow-outdated-plugins");
    * </code></pre>
@@ -125,6 +133,20 @@ public class ChromiumOptions<T extends ChromiumOptions<?>> extends AbstractDrive
    * @param arguments The arguments to use when starting Chrome.
    */
   public T addArguments(List<String> arguments) {
+    /*
+      --remote-allow-origins is being added by default since Chrome 111. We need to check
+      if the argument already exists and then remove it.
+     */
+    String remoteAllowOrigins = "remote-allow-origins";
+    Optional<String> newArg = arguments.stream()
+      .filter(arg -> arg.contains(remoteAllowOrigins))
+      .findFirst();
+    Optional<String> existingArg = args.stream()
+      .filter(arg -> arg.contains(remoteAllowOrigins))
+      .findFirst();
+    if (newArg.isPresent() && existingArg.isPresent()) {
+      args.remove(existingArg.get());
+    }
     args.addAll(arguments);
     return (T) this;
   }