+++ java/src/org/openqa/selenium/firefox/FirefoxDriver.java
@@ -124,20 +124,29 @@ public class FirefoxDriver extends RemoteWebDriver
   }
 
   public FirefoxDriver(FirefoxDriverService service, FirefoxOptions options) {
-    this(generateExecutor(service, options), options);
+    this(service, options, ClientConfig.defaultConfig());
   }
 
-  private static FirefoxDriverCommandExecutor generateExecutor(FirefoxDriverService service, FirefoxOptions options) {
+  public FirefoxDriver(FirefoxDriverService service, FirefoxOptions options, ClientConfig clientConfig) {
+    this(generateExecutor(service, options, clientConfig), options);
+  }
+
+  private static FirefoxDriverCommandExecutor generateExecutor(FirefoxDriverService service, FirefoxOptions options, ClientConfig clientConfig) {
     Require.nonNull("Driver service", service);
     Require.nonNull("Driver options", options);
+    Require.nonNull("Driver clientConfig", clientConfig);
     if (service.getExecutable() == null) {
       String path = DriverFinder.getPath(service, options);
       service.setExecutable(path);
     }
-    return new FirefoxDriverCommandExecutor(service);
+    return new FirefoxDriverCommandExecutor(service, clientConfig);
   }
 
   private FirefoxDriver(FirefoxDriverCommandExecutor executor, FirefoxOptions options) {
+    this(executor, options, ClientConfig.defaultConfig());
+  }
+
+  private FirefoxDriver(FirefoxDriverCommandExecutor executor, FirefoxOptions options, ClientConfig clientConfig) {
     super(executor, checkCapabilitiesAndProxy(options));
     webStorage = new RemoteWebStorage(getExecuteMethod());
     extensions = new AddHasExtensions().getImplementation(getCapabilities(), getExecuteMethod());
@@ -162,10 +171,10 @@ public class FirefoxDriver extends RemoteWebDriver
 
     this.cdpUri = cdpUri;
     this.capabilities = cdpUri.map(uri ->
-                                     new ImmutableCapabilities(
-                                       new PersistentCapabilities(capabilities)
-                                         .setCapability("se:cdp", uri.toString())
-                                         .setCapability("se:cdpVersion", "85.0")))
+        new ImmutableCapabilities(
+          new PersistentCapabilities(capabilities)
+            .setCapability("se:cdp", uri.toString())
+            .setCapability("se:cdpVersion", "85.0")))
       .orElse(new ImmutableCapabilities(capabilities));
   }
 
@@ -202,7 +211,7 @@ public class FirefoxDriver extends RemoteWebDriver
   public void setFileDetector(FileDetector detector) {
     throw new WebDriverException(
       "Setting the file detector only works on remote webdriver instances obtained " +
-      "via RemoteWebDriver");
+        "via RemoteWebDriver");
   }
 
   @Override
@@ -260,6 +269,10 @@ public class FirefoxDriver extends RemoteWebDriver
     context.setContext(commandContext);
   }
 
+  /**
+   * @deprecated Use W3C-compliant BiDi protocol. Use {{@link #maybeGetBiDi()}}
+   */
+  @Deprecated
   @Override
   public Optional<DevTools> maybeGetDevTools() {
     if (devTools != null) {
@@ -284,6 +297,10 @@ public class FirefoxDriver extends RemoteWebDriver
     return Optional.of(devTools);
   }
 
+  /**
+   * @deprecated Use W3C-compliant BiDi protocol. Use {{@link #getBiDi()}}
+   */
+  @Deprecated
   @Override
   public DevTools getDevTools() {
     if (!cdpUri.isPresent()) {
@@ -343,7 +360,9 @@ public class FirefoxDriver extends RemoteWebDriver
 
     /**
      * System property that defines the location of the file where Firefox log should be stored.
+     * @deprecated equivalent constant located at {@link GeckoDriverService#GECKO_DRIVER_LOG_PROPERTY}
      */
+    @Deprecated
     public static final String BROWSER_LOGFILE = "webdriver.firefox.logfile";
 
     /**
@@ -357,7 +376,11 @@ public class FirefoxDriver extends RemoteWebDriver
   private static class FirefoxDriverCommandExecutor extends DriverCommandExecutor {
 
     public FirefoxDriverCommandExecutor(DriverService service) {
-      super(service, getExtraCommands());
+      this(service, ClientConfig.defaultConfig());
+    }
+
+    public FirefoxDriverCommandExecutor(DriverService service, ClientConfig clientConfig) {
+      super(service, getExtraCommands(), clientConfig);
     }
 
     private static Map<String, CommandInfo> getExtraCommands() {